+++ polaris-core/src/main/java/org/apache/polaris/core/entity/PolarisEntityConstants.java
@@ -43,6 +43,9 @@ public class PolarisEntityConstants {
   // the name of the principal role we create to manage the entire Polaris service
   private static final String ADMIN_PRINCIPAL_ROLE_NAME = "service_admin";
 
+  // the name of the principal role for catalog admins to list principal roles
+  private static final String PRINCIPAL_ROLE_VIEWER_ROLE_NAME = "principal_role_viewer";
+
   // 24 hours retention before purging. This should be a config
   private static final long RETENTION_TIME_IN_MS = 24 * 3600_000;
 
@@ -89,6 +92,10 @@ public class PolarisEntityConstants {
 
   public static String getNameOfPrincipalServiceAdminRole() {
     return ADMIN_PRINCIPAL_ROLE_NAME;
+  }
+
+  public static String getNameOfPrincipalRoleViewerRole() {
+    return PRINCIPAL_ROLE_VIEWER_ROLE_NAME;
   }
 
   public static long getRetentionTimeInMs() {